크기가 작은 부분 문자열 문제 풀이

+++ [PGM] Small Portion Strings/ParkHyunS00.java
@@ -21,11 +21,11 @@ class Solution1 {
 class Solution2 {
     public int solution(String t, String p) {
         int answer = 0;
-        int pNum = Integer.parseInt(p);
+        long pNum = Long.parseLong(p);
 
         for(int i=0; i<t.length() - p.length() + 1; i++){
             String buff = t.substring(i, i + p.length());
-            int compareNum = Integer.parseInt(buff);
+            long compareNum = Long.parseLong(buff);
 
             if (compareNum <= pNum) ++answer;
         }
